Added I2C module files
