+++ PSLab_Original/Common_Methods.h
@@ -13,6 +13,13 @@
 #include <libpic30.h>
 
 #define DELAY_105uS asm volatile ("REPEAT, #6721"); Nop(); // 105uS delay
+#define ERROR_BUFFLEN 1500
+
+/*--------Error handling definitions------*/
+static char errors[ERROR_BUFFLEN];
+static char *error_readpos, *error_writepos;
+
+void logit(char *str);
 
 void Delay_us(unsigned int delay);
 void Delay_us_by8(unsigned int delay);
