+++ PSLab_Original/nbproject/private/private.xml
@@ -4,7 +4,13 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
+            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/Common_Methods.h</file>
+            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/PSLAB_I2C.c</file>
+            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/PSLAB_I2C.h</file>
+            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/functions.c</file>
+            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/functions.h</file>
             <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/proto2_main.c</file>
+            <file>file:/home/padmal/Documents/FossAsia/pslab-firmware/PSLab_Original/Common_Methods.c</file>
         </group>
     </open-files>
 </project-private>
